refactor: Update calculator.py to convert user input to floats before calculating the sum

+++ 00-Functions-Variables/calculator.py
@@ -1,6 +1,6 @@
 # Calculator
 
-x = int(input("What's x? "))
-y = int(input("What's y? "))
+x = float(input("What's x? "))
+y = float(input("What's y? "))
 
 print(x + y)
